中间平台卖品接口

+++ src/main/java/com/boot/security/server/api/ctms/reply/Dy1905ModelMapper.java
@@ -94,7 +94,7 @@ public class Dy1905ModelMapper {
     	entity.setGoodsCode(model.getSerial());
     	entity.setGoodsName(model.getName());
     	if(model.getPrice()!=null){
-    		entity.setStandardPrice(BigDecimal.valueOf(Double.valueOf(model.getPrice())));
+    		entity.setStandardPrice((Double.valueOf(model.getPrice())));
     	}
     	entity.setGoodsPic(model.getImage());
     	entity.setStockCount(model.getStockCount() ==null?0 : Integer.valueOf(model.getStockCount()));
